regression and classification up and running

+++ scripts/utils.py
@@ -2,6 +2,7 @@ import itertools
 import os
 from subprocess import call
 
+
 def make_experiment_combinations(combinations : list):
     """
     The product of all combinations of arguments.
@@ -26,17 +27,21 @@ def make_experiment_combinations(combinations : list):
 
     return ret
 
-def make_local_jobs(script : str, experiments : list):
+def make_local_jobs(script : str, experiments : list, overwrite=False):
     """
     Writes a file of scripts to be run in series on a single machine.
+    
+    If overwrite=True then a new file is written with a shebang, otherwise lines are appended. 
 
     :param script: name of python script to run
     :param experiments: list of dictionaries of args
     :return: None
     """
-    with open('local_run', 'w') as f:
-        f.write('#!/usr/bin/env bash\n\n')
+    if overwrite:
+        with open('local_run', 'w') as f:
+            f.write('#!/usr/bin/env bash\n\n')
 
+    with open('local_run', 'a') as f:
         for e in experiments:
             s = 'python {}.py '.format(script)
             for k in e:
@@ -44,17 +49,21 @@ def make_local_jobs(script : str, experiments : list):
             s += '\n'
             f.write(s)
 
-def make_condor_jobs(script : str, experiments : list):
+def make_condor_jobs(script : str, experiments : list, overwrite=False):
     """
-    Writes a very basic condor submission file, and also creates the executable if necessary. Preamble for the 
-    exectable (e.g. for setting up the python environment) should go in 'preamble.txt.txt',  
+    Writes a condor submission file, and also creates the executable if necessary. Preamble for the 
+    exectable (e.g. for setting up the python environment) should go in 'preamble.txt.txt'. Preamble
+    for the condor submission should go in condor_preamble.txt.txt.
+    
+    If overwrite=True then a new file is written with the condor preamble from condor_preamble.txt,
+    otherwise lines are appended. 
 
     :param script: name of python script to run
     :param experiments: list of dictionaries of args
     :return: None
     """
     condor_run_file = 'condor_run'
-    if True:# not os.path.isfile(condor_run_file):
+    if not os.path.isfile(condor_run_file):
         with open(condor_run_file, 'w') as f:
             f.write("#!/usr/bin/env bash\n")
 
@@ -73,23 +82,12 @@ def make_condor_jobs(script : str, experiments : list):
 
         call(["chmod", '777', condor_run_file])
 
+    if overwrite:
+        with open('condor_jobs', 'w') as f:
+            with open('condor_preamble.txt', 'r') as ff:
+                f.writelines(ff)
 
-    with open('condor_jobs', 'w') as f:
-        f.write('Universe   = vanilla\n')
-        f.write('Executable = condor_run \n')
-
-        # or send these elsewhere
-        # f.write('Log        = /dev/null\n')
-        # f.write('Output     = /dev/null\n')
-        # f.write('Error      = /dev/null\n')
-
-        f.write('Log        = /vol/bitbucket/hrs13/condor/$(Process)_log.txt\n')
-        f.write('Output     = /vol/bitbucket/hrs13/condor/$(Process)_stdout.txt\n')
-        f.write('Error      = /vol/bitbucket/hrs13/condor/$(Process)_stderr.txt\n')
-
-
-        f.write('\n')
-
+    with open('condor_jobs', 'a') as f:
         for e in experiments:
             t = 'Arguments = {}.py '.format(script)
             for k in e:
